refactor: extract usd value logic from SwapTab
